+++ app/Http/Controllers/API/UsersController.php
@@ -6,11 +6,22 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use App\User;
+use App\RolePermission;
 use App\Http\Resources\User as UserResource;
 use App\Http\Resources\UserCollection;
 
 class UsersController extends Controller
 {
+    public function Permission ()
+    {
+        $usersRole = auth('api')->user()->role_id;
+        $permission = RolePermission::
+            where('object', "Users")->
+            where('role_id', $usersRole)->
+            first();
+        return ($permission);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -18,6 +29,9 @@ class UsersController extends Controller
      */
     public function index(Request $request)
     {
+        if (!$this->Permission()->b)
+            return \Response::json(['error' => 'Not nought privileges' ], 401);
+
         $users = User::all();
         return new UserCollection($users);
     }
@@ -30,6 +44,9 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
+        if (!$this->Permission()->a)
+            return \Response::json(['error' => 'Not nought privileges' ], 401);
+
         if (is_null($request->password)) {
             $request->request->add(['password'=>\Illuminate\Support\Str::random(32)]);
         }
@@ -66,6 +83,9 @@ class UsersController extends Controller
      */
     public function update(Request $request)
     {
+        if (!$this->Permission()->e)
+            return \Response::json(['error' => 'Not nought privileges' ], 401);
+
         $request->validate([
                 'id'            => 'required|integer',
                 'name'          => 'required|max:255',
@@ -93,6 +113,9 @@ class UsersController extends Controller
      */
     public function show($id)
     {
+        if (!$this->Permission()->r)
+            return \Response::json(['error' => 'Not nought privileges' ], 401);
+
         $user = User::findOrFail($id);
 
         if ($user!=null) {
@@ -110,12 +133,15 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
+        if (!$this->Permission()->d)
+            return \Response::json(['error' => 'Not nought privileges' ], 401);
+
         $user = User::findOrFail($id);
 
         if ($user->delete()) {
             return new UserResource($user);
         } else {
-            return Response::json(['error' => 'Resource not found' ], 404);
+            return \Response::json(['error' => 'Resource not found' ], 404);
         }
     }
 }
